requirements.txt
PyQt5
pillow
numpy
pytest-shutil
matplotlib